+++ sdk/trace/span.go
@@ -497,8 +497,8 @@ func (s *recordingSpan) End(options ...trace.SpanEndOption) {
 	}
 	s.mu.Unlock()
 
-	defer func() {
-		if s.tracer.selfObservabilityEnabled {
+	if s.tracer.selfObservabilityEnabled {
+		defer func() {
 			// Determine the sampling result and create the corresponding attribute.
 			var attrSamplingResult attribute.KeyValue
 			if s.spanContext.IsSampled() {
@@ -513,8 +513,8 @@ func (s *recordingSpan) End(options ...trace.SpanEndOption) {
 			// with the correct span context.
 			ctx := trace.ContextWithSpan(context.Background(), s)
 			s.tracer.spanLiveMetric.Add(ctx, -1, attrSamplingResult)
-		}
-	}()
+		}()
+	}
 
 	sps := s.tracer.provider.getSpanProcessors()
 	if len(sps) == 0 {