+++ sdk/trace/trace_test.go
@@ -2679,7 +2679,8 @@ func TestSelfObservability(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Setenv("OTEL_GO_X_SELF_OBSERVABILITY", "True")
 			prev := otel.GetMeterProvider()
-			defer otel.SetMeterProvider(prev)
+			t.Cleanup(func() { otel.SetMeterProvider(prev) })
+
 			p := &producer{}
 			r := metric.NewManualReader(metric.WithProducer(p))
 			mp := metric.NewMeterProvider(metric.WithReader(r))