+++ CHANGELOG.md
@@ -71,6 +71,7 @@ The next release will require at least [Go 1.24].
 ### Fixed
 
 - Fix `go.opentelemetry.io/otel/exporters/prometheus` to deduplicate suffixes if already present in metric name when UTF8 is enabled. (#7088)
+- Fix the `go.opentelemetry.io/otel/exporters/stdout/stdouttrace` self-observability component type and name. (#7195)
 - Propagate context to self-observability measurements in `go.opentelemetry.io/otel/sdk/trace`. (#7209)
 
 <!-- Released section -->